[misc] Fixing the migration of StaticListClassFieldTest to a docker integration test
- Add testMultipleSelect inherited from AbstractListClassFieldTest to StaticListClassFieldIT
- Move isReadOnly up to ListClassFieldEditPane
- Move testMultipleSelect to DBListClassFieldTest
- Use ListClassFieldEditPane#isReadOnly in DBListClassFieldTest#testMultipleSelect
- Remove the now unused AbstractListClassFieldTest
