+++ xwiki-platform-core/xwiki-platform-appwithinminutes/xwiki-platform-appwithinminutes-test/xwiki-platform-appwithinminutes-test-pageobjects/src/main/java/org/xwiki/appwithinminutes/test/po/ListClassFieldEditPane.java
@@ -160,4 +160,14 @@ public class ListClassFieldEditPane extends SuggestClassFieldEditPane
         }
         return null;
     }
+
+    /**
+     * @return {@code true} if the field is readonly, {@code false} otherwise
+     * @since 13.2
+     * @since 12.10.6
+     */
+    public boolean isReadOnly()
+    {
+        return getSizeInput().getAttribute("readOnly") != null;
+    }
 }
