+++ xwiki-platform-core/xwiki-platform-appwithinminutes/xwiki-platform-appwithinminutes-test/xwiki-platform-appwithinminutes-test-pageobjects/src/main/java/org/xwiki/appwithinminutes/test/po/StaticListClassFieldEditPane.java
@@ -47,14 +47,4 @@ public class StaticListClassFieldEditPane extends ListClassFieldEditPane
         return new StaticListItemsEditor(
             getDriver().findElementWithoutWaiting(getContainer(), By.className("staticListEditor")));
     }
-
-    /**
-     * @return {@code true} if the field is readonly, {@code false} otherwise
-     * @since 13.2
-     * @since 12.10.6
-     */
-    public boolean isReadOnly()
-    {
-        return getSizeInput().getAttribute("readOnly") != null;
-    }
 }
